requirements.txt
Adafruit-CharLCD==1.0.0
Adafruit-GPIO==0.9.3
amqp==1.4.9
anyjson==0.3.3
argh==0.26.1
async==0.6.2
Babel==1.3
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.18
biplist==0.5
blaze==0.7.2
blinker==1.4
blosc==1.2.4
boto==2.34.0
bottle==0.12.7
Bottleneck==1.0.0
Brownie==0.5.1
cairocffi==0.5.4
celery==3.1.5
celerymon==1.0.3
certifi==2015.11.20.1
cffi==1.4.2
chardet==2.1.1
Cheetah==2.4.4
CherryPy==3.6.0
ClientForm==0.2.10
configobj==5.0.6
Couchapp==1.0.1
CouchDB==1.0
couchdbkit==0.6.5
couchquery==0.9
cov-core==1.13.0
coverage==3.7.1
cryptography==1.3.1
cssselect==0.9.1
Cython==0.21
cytoolz==0.7.2
datashape==0.4.5
debtcollector==1.2.0
decorator==3.4.2
Distutils2==1.0a4
Django==1.6.11
django-appconf==0.6
django-athumb==2.0
django-auth-ldap==1.2.0
django-authopenid==1.0.1
django-celery==3.0.23
django-compressor==1.4
django-dajax==0.9.2
django-dajaxice==0.5.5
django-dbtemplates==1.3
django-debug-toolbar==1.2.1
django-grappelli==2.4.6
django-guardian==1.1.1
django-haystack==2.4.0rc1
django-international==0.0.6
django-mailer==0.1.0
django-mediasync==2.2.0
django-navigation==0.8.1
django-nose==1.2
django-nose-selenium==0.7.3
django-openstack-auth==1.1.7
django-picklefield==0.3.0
django-piston==0.2.3
django-registration==1.0
django-static==1.5.5
django-tastypie==0.11.1
dnspython==1.11.1
docopt==0.6.2
docutils==0.12
dogpile.cache==0.5.4
dogpile.core==0.4.1
dojango==0.5.5
dtopt==0.1
dulwich==0.9.3
EasyProcess==0.2.1
ecdsa==0.13
egenix-mx-base==3.2.8
enum-compat==0.0.2
enum34==1.1.3
epydoc==3.0.1
eventlet==0.17.4
extras==0.0.3
flake8==2.2.2
Flask==0.10.1
Flask-API==0.6.4
Flask-API-Utils==1.0.2
Flask-Auth==0.85
Flask-Babel==0.9
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
funcsigs==0.4
futures==2.1.6
Genshi==0.7
gevent==1.0.2
gitdb==0.5.4
GitPython==0.3.2rc1
gmixer==1.0
google-api-python-client==1.2
greenlet==0.4.7
guess-language==0.2
h5py==2.5.0
haystack==0.15
hdf5storage==0.1.1
hg-git==0.6.1
hgtools==6.0
html5lib==0.999
http-parser==0.8.3
httplib2==0.9.2
idna==2.0
image==1.4.0
into==0.2.2
ipaddr==2.1.11
ipaddress==1.0.14
ipython==2.2.0
iso8601==0.1.10
itsdangerous==0.24
java-gcj-compat==1.0.80
javapackages==4.6.0
jedi===0.8.1-final0
Jinja2==2.7.3
jsonschema==2.4.0
keyring==9.0
keystoneauth1==2.2.0
Kivy-Garden==0.1.1
kombu==3.0.33
lazyarray==0.2.2
libvirt-python==1.3.3
logutils==0.3.3
lxml==3.5.0
Markdown==2.6.2
MarkupSafe==0.23
matplotlib==1.4.3
mccabe==0.2.1
mechanize==0.2.5
mercurial==3.1.2
mock==1.0.1
monotonic==0.6
msgpack-python==0.4.7
multipledispatch==0.4.7
mysql-connector-python==1.2.3
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
netaddr==0.7.12
netifaces==0.10.4
networkx==1.9
NeuroTools==0.1.0
nltk==3.0.2
nose==1.3.3
Numdifftools==0.4.0
numexpr==2.4
numpy==1.11.0
numpydoc==0.5
nxt-python==2.2.2
oauth==1.0.1
oauthlib==0.6.3
odict==1.5.1
openpyxl==1.7.0
oslo.config==1.4.0
oslo.i18n==3.2.0
oslo.serialization==2.2.0
oslo.utils==3.4.0
oslo.vmware==0.5.0
pandas==0.16.1
paramiko==1.16.0
pathtools==0.1.2
patsy==0.2.1
pbr==0.10.0
pep8==1.6.2
pexpect==3.3
Pillow==2.8.1
ply==3.4
prettytable==0.7.2
protobuf==2.5.0
psutil==2.2.1
psycopg2==2.6
py==1.4.26
pyalsaaudio==0.7
pyasn1==0.1.9
pyasn1-modules==0.0.5
PyAudio==0.2.8
pycares==0.6.1
pychecker==0.8.19
pycparser==2.14
pycrypto==2.6.1
pycups==1.9.72
pycurl==7.19.5.1
pydebug==1.0.3
pyenchant==1.6.5
pyFFTW==0.9.2
PyFFTW3==0.2.1
pyflakes==0.8.1
Pygments==1.6
pygobject==3.20.1
pygraphviz==1.2
PyGreSQL==4.1.1
pyinotify==0.9.4
PyJWT==0.2.1
pymatbridge==0.4.3
pymc==2.3
pymongo==2.7.2
PyMySQL==0.6
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pyOpenSSL==0.15.1
PyPAM==0.5.0
pyparsing==2.0.1
pyplete==0.0.5
pyqtgraph==0.9.10
pyrtlsdr==0.2.0
pyserial==2.7
pysmell==0.7.3
pysolr==3.3.2
PySoundCard==0.5.0
PySoundFile==0.5.0
pysqlite==2.6.3
pytest==2.7.0
pytest-cov==1.7.0
pytest-runner==2.0
python-dateutil==2.1
python-digest==1.7
python-keystoneclient==0.11.0
python-ldap==2.4.16
python-mimeparse==0.1.4
python-openid==2.2.5
python-ptrace==0.8.1
python-xlib==0.12
pytz==2014.4
pyudev==0.17
pyusb==1.0.0b1
PyVirtualDisplay==0.1.5
PyXB==1.2.4
pyxdg==0.25
PyYAML==3.11
pyzmq==14.0.0
quantities==0.10.1
queuelib==1.1.1
requests==2.9.1
restkit==4.2.2
RPi.GPIO==0.5.11
scapy==2.3.2
scikit-learn==0.14.1
scipy==0.17.0
Scrapy==0.24.4
selenium==2.44.0
setproctitle==1.1.8
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.10.0
slimmer==0.1.30
smmap==0.8.2
socketpool==0.5.3
speaklater==1.3
Sphinx==1.2.3
spidev==3.1
SQLAlchemy==0.9.7
sqlalchemy-migrate==0.10.0
sqlparse==0.1.10
statsmodels==0.6.0
stevedore==1.12.0
suds==0.4
Tempita==0.5.2
toolz==0.7.2
tornado==3.2.1
Twisted==15.5.0
uritemplate==0.6
urlgrabber==3.9.1
urwid==1.2.1
vboxapi==1.0
vext==0.4.99.5
vext.pygame==0.4.2
virtualenv==13.1.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
w3lib==1.10.0
watchdog==0.6.0
weave==0.15.0
WebOb==1.4
websockify==0.6.1
Werkzeug==0.9.6
Whoosh==2.7.0
wrapt==1.10.6
WTForms==2.0.2
wxPython-common==3.0.2.0
xlrd==0.9.2
xlwt==0.7.5
xmldiff==0.6.10
xmltodict==0.9.2
zeroconf==0.17.4
zope.interface==4.1.2
